// src/api/client/models/translate-text.ts
/* tslint:disable */
/* eslint-disable */
/**
 * booknest
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TranslateText
 */
export interface TranslateText {
    /**
     * 
     * @type {string}
     * @memberof TranslateText
     */
    'targetLang': string;
    /**
     * 
     * @type {string}
     * @memberof TranslateText
     */
    'text': string;
}